student-module/requirements.txt
mediapipe==0.10.8
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
tensorflow==2.15.0
pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6 